refactor schedule.ts into schedules dir
